+++ uberenv.py
@@ -1,7 +1,7 @@
 #!/bin/sh
 "exec" "python" "-u" "-B" "$0" "$@"
 ###############################################################################
-# Copyright (c) 2014-2020, Lawrence Livermore National Security, LLC.
+# Copyright (c) 2014-2021, Lawrence Livermore National Security, LLC.
 #
 # Produced at the Lawrence Livermore National Laboratory
 #
@@ -47,7 +47,8 @@
 """
  file: uberenv.py
 
- description: automates using spack to install a project.
+ description: automates using a package manager to install a project.
+ Uses spack on Unix-based systems and Vcpkg on Windows-based systems.
 
 """
 
@@ -72,7 +73,7 @@ from os.path import abspath as pabs
 def sexe(cmd,ret_output=False,echo=False):
     """ Helper for executing shell commands. """
     if echo:
-        print("[exe: {}]".format(cmd))
+        print("[exe: {0}]".format(cmd))
     if ret_output:
         p = subprocess.Popen(cmd,
                              shell=True,
@@ -105,6 +106,12 @@ def parse_args():
                       dest="spec",
                       default=None,
                       help="spack compiler spec")
+
+    # for vcpkg, what architecture to target
+    parser.add_option("--triplet",
+                      dest="vcpkg_triplet",
+                      default=None,
+                      help="vcpkg architecture triplet")
 
     # optional location of spack mirror
     parser.add_option("--mirror",
@@ -164,12 +171,12 @@ def parse_args():
                       default=False,
                       help="Ignore SSL Errors")
 
-    # option to force a spack pull
+    # option to force a pull of the package manager
     parser.add_option("--pull",
                       action="store_true",
-                      dest="spack_pull",
+                      dest="repo_pull",
                       default=False,
-                      help="Pull if spack repo already exists")
+                      help="Pull from package manager, if repo already exists")
 
     # option to force for clean of packages specified to
     # be cleaned in the project.json
@@ -214,7 +221,7 @@ def parse_args():
     if not opts["spack_config_dir"] is None:
         opts["spack_config_dir"] = pabs(opts["spack_config_dir"])
         if not os.path.isdir(opts["spack_config_dir"]):
-            print("[ERROR: invalid spack config dir: {} ]".format(opts["spack_config_dir"]))
+            print("[ERROR: invalid spack config dir: {0} ]".format(opts["spack_config_dir"]))
             sys.exit(-1)
     # if rel path is given for the mirror, we need to evaluate here -- before any
     # chdirs to avoid confusion related to what it is relative to.
@@ -275,6 +282,9 @@ class UberEnv():
         # load project settings
         self.project_opts = load_json_file(opts["project_json"])
 
+        # setup main package name
+        self.pkg_name = self.set_from_args_or_json("package_name")
+
         # Set project.json defaults
         if not "force_commandline_prefix" in self.project_opts:
             self.project_opts["force_commandline_prefix"] = False
@@ -290,11 +300,28 @@ class UberEnv():
     def setup_paths_and_dirs(self):
         self.uberenv_path = uberenv_script_dir()
 
+        # setup destination paths
+        if not self.opts["prefix"]:
+            if self.project_opts["force_commandline_prefix"]:
+                # project has specified prefix must be on command line
+                print("[ERROR: --prefix flag for library destination is required]")
+                sys.exit(1)
+            # otherwise set default
+            self.opts["prefix"] = "uberenv_libs"
+
+        self.dest_dir = pabs(self.opts["prefix"])
+
+        # print a warning if the dest path already exists
+        if not os.path.isdir(self.dest_dir):
+            os.mkdir(self.dest_dir)
+        else:
+            print("[info: destination '{0}' already exists]".format(self.dest_dir))
+
     def set_from_args_or_json(self,setting):
         try:
             setting_value = self.project_opts[setting]
         except (KeyError):
-            print("ERROR: {} must at least be defined in project.json".format(setting))
+            print("ERROR: '{0}' must at least be defined in project.json".format(setting))
             raise
         else:
             if self.opts[setting]:
@@ -305,7 +332,7 @@ class UberEnv():
         try:
             setting_value = self.project_opts[setting]
         except (KeyError):
-            print("ERROR: {} must at least be defined in project.json".format(setting))
+            print("ERROR: '{0}' must at least be defined in project.json".format(setting))
             raise
         return setting_value
 
@@ -320,13 +347,127 @@ class UberEnv():
         return res
 
 
+class VcpkgEnv(UberEnv):
+    """ Helper to clone vcpkg and install libraries on Windows """
+
+    def __init__(self, opts, extra_opts):
+        UberEnv.__init__(self,opts,extra_opts)
+
+        # setup architecture triplet
+        self.vcpkg_triplet = self.set_from_args_or_json("vcpkg_triplet")
+        print("Vcpkg triplet: {}".format(self.vcpkg_triplet))
+        if self.vcpkg_triplet is None:
+           self.vcpkg_triplet = os.getenv("VCPKG_DEFAULT_TRIPLET", "x86-windows")
+
+    def setup_paths_and_dirs(self):
+        # get the current working path, and the glob used to identify the
+        # package files we want to hot-copy to vcpkg
+
+        UberEnv.setup_paths_and_dirs(self)
+
+        self.ports = pjoin(self.uberenv_path, "vcpkg_ports","*")
+
+        # setup path for vcpkg repo
+        print("[installing to: {0}]".format(self.dest_dir))
+        self.dest_vcpkg = pjoin(self.dest_dir,"vcpkg")
+
+        if os.path.isdir(self.dest_vcpkg):
+            print("[info: destination '{0}' already exists]".format(self.dest_vcpkg))
+
+    def clone_repo(self):
+        if not os.path.isdir(self.dest_vcpkg):
+            # compose clone command for the dest path, vcpkg url and branch
+            vcpkg_branch = self.project_opts.get("vcpkg_branch", "master")
+            vcpkg_url = self.project_opts.get("vcpkg_url", "https://github.com/microsoft/vcpkg")
+
+            print("[info: cloning vcpkg '{0}' branch from {1} into {2}]"
+                .format(vcpkg_branch,vcpkg_url, self.dest_vcpkg))
+
+            os.chdir(self.dest_dir)
+
+            clone_opts = ("-c http.sslVerify=false " 
+                          if self.opts["ignore_ssl_errors"] else "")
+
+            clone_cmd =  "git {0} clone --single-branch --depth=1 -b {1} {2} vcpkg".format(clone_opts, vcpkg_branch,vcpkg_url)
+            sexe(clone_cmd, echo=True)
+
+            # optionally, check out a specific commit
+            if "vcpkg_commit" in self.project_opts:
+                sha1 = self.project_opts["vcpkg_commit"]
+                print("[info: using vcpkg commit {0}]".format(sha1))
+                os.chdir(self.dest_vcpkg)
+                sexe("git checkout {0}".format(sha1),echo=True)
+                
+        if self.opts["repo_pull"]:
+            # do a pull to make sure we have the latest
+            os.chdir(self.dest_vcpkg)
+            sexe("git stash", echo=True)
+            res = sexe("git pull", echo=True)
+            if res != 0:
+                #Usually untracked files that would be overwritten
+                print("[ERROR: Git failed to pull]")
+                sys.exit(-1)
+
+
+        # Bootstrap vcpkg
+        os.chdir(self.dest_vcpkg)
+        print("[info: bootstrapping vcpkg]")
+        sexe("bootstrap-vcpkg.bat -disableMetrics")
+
+    def patch(self):
+        """ hot-copy our ports into vcpkg """
+        
+        import distutils.dir_util
+
+        src_vcpkg_ports = pjoin(self.uberenv_path, "vcpkg_ports")
+        dest_vcpkg_ports = pjoin(self.dest_vcpkg,"ports")
+
+        print("[info: copying from {0} to {1}]".format(src_vcpkg_ports,dest_vcpkg_ports))
+        distutils.dir_util.copy_tree(src_vcpkg_ports,dest_vcpkg_ports)
+
+
+    def clean_build(self):
+        pass
+
+    def show_info(self):
+        os.chdir(self.dest_vcpkg)
+        print("[info: Details for package '{0}']".format(self.pkg_name))
+        sexe("vcpkg.exe search " + self.pkg_name, echo=True)
+
+        print("[info: Dependencies for package '{0}']".format(self.pkg_name))
+        sexe("vcpkg.exe depend-info " + self.pkg_name, echo=True)
+
+    def create_mirror(self):
+        pass
+
+    def use_mirror(self):
+        pass
+
+    def install(self):
+        
+        os.chdir(self.dest_vcpkg)
+        install_cmd = "vcpkg.exe "
+        install_cmd += "install {0}:{1}".format(self.pkg_name, self.vcpkg_triplet)
+
+        res = sexe(install_cmd, echo=True)
+
+        # Running the install_cmd eventually generates the host config file,
+        # which we copy to the target directory.
+        src_hc = pjoin(self.dest_vcpkg, "installed", self.vcpkg_triplet, "include", self.pkg_name, "hc.cmake")
+        hcfg_fname = pjoin(self.dest_dir, "{0}.{1}.cmake".format(platform.uname()[1], self.vcpkg_triplet))
+        print("[info: copying host config file to {0}]".format(hcfg_fname))
+        shutil.copy(os.path.abspath(src_hc), hcfg_fname)
+        print("")
+        print("[install complete!]")
+        return res
+
+
 class SpackEnv(UberEnv):
     """ Helper to clone spack and install libraries on MacOS an Linux """
 
     def __init__(self, opts, extra_opts):
         UberEnv.__init__(self,opts,extra_opts)
 
-        self.pkg_name = self.set_from_args_or_json("package_name")
         self.pkg_version = self.set_from_json("package_version")
         self.pkg_final_phase = self.set_from_args_or_json("package_final_phase")
         self.pkg_src_dir = self.set_from_args_or_json("package_source_dir")
@@ -350,13 +491,13 @@ class SpackEnv(UberEnv):
                 opts["spec"] = "%clang"
             else:
                 opts["spec"] = "%gcc"
-            self.opts["spec"] = "@{}{}".format(self.pkg_version,opts["spec"])
+            self.opts["spec"] = "@{0}{1}".format(self.pkg_version,opts["spec"])
         elif not opts["spec"].startswith("@"):
-            self.opts["spec"] = "@{}{}".format(self.pkg_version,opts["spec"])
+            self.opts["spec"] = "@{0}{1}".format(self.pkg_version,opts["spec"])
         else:
-            self.opts["spec"] = "{}".format(opts["spec"])
+            self.opts["spec"] = "{0}".format(opts["spec"])
 
-        print("[spack spec: {}]".format(self.opts["spec"]))
+        print("[spack spec: {0}]".format(self.opts["spec"]))
 
 
     def append_path_to_packages_paths(self, path, errorOnNonexistant=True):
@@ -404,40 +545,24 @@ class SpackEnv(UberEnv):
             # default to packages living next to uberenv script if it exists
             self.append_path_to_packages_paths(pjoin(self.uberenv_path,"packages"), errorOnNonexistant=False)
 
-        # setup destination paths
-        if not self.opts["prefix"]:
-            if self.project_opts["force_commandline_prefix"]:
-                # project has specified prefix must be on command line
-                print("[ERROR: --prefix flag for library destination is required]")
-                sys.exit(1)
-            # otherwise set default
-            self.opts["prefix"] = "uberenv_libs"
-            
-        self.dest_dir = pabs(self.opts["prefix"])
-        self.dest_spack = pjoin(self.dest_dir,"spack")
         print("[installing to: {0}]".format(self.dest_dir))
 
-        # print a warning if the dest path already exists
-        if not os.path.isdir(self.dest_dir):
-            os.mkdir(self.dest_dir)
-        else:
-            print("[info: destination '{}' already exists]".format(self.dest_dir))
-
+        self.dest_spack = pjoin(self.dest_dir,"spack")
         if os.path.isdir(self.dest_spack):
-            print("[info: destination '{}' already exists]".format(self.dest_spack))
+            print("[info: destination '{0}' already exists]".format(self.dest_spack))
 
         self.pkg_src_dir = os.path.join(self.uberenv_path,self.pkg_src_dir)
         if not os.path.isdir(self.pkg_src_dir):
-            print("[ERROR: package_source_dir '{}' does not exist]".format(self.pkg_src_dir))
+            print("[ERROR: package_source_dir '{0}' does not exist]".format(self.pkg_src_dir))
             sys.exit(-1)
 
 
     def find_spack_pkg_path_from_hash(self, pkg_name, pkg_hash):
-        res, out = sexe("spack/bin/spack find -p /{}".format(pkg_hash), ret_output = True)
+        res, out = sexe("spack/bin/spack find -p /{0}".format(pkg_hash), ret_output = True)
         for l in out.split("\n"):
             if l.startswith(pkg_name):
                    return {"name": pkg_name, "path": l.split()[-1]}
-        print("[ERROR: failed to find package named '{}']".format(pkg_name))
+        print("[ERROR: failed to find package named '{0}']".format(pkg_name))
         sys.exit(-1)
 
     def find_spack_pkg_path(self, pkg_name, spec = ""):
@@ -447,7 +572,7 @@ class SpackEnv(UberEnv):
             # pick the first in the list.
             if l.startswith(pkg_name):
                    return {"name": pkg_name, "path": l.split()[-1]}
-        print("[ERROR: failed to find package named '{}']".format(pkg_name))
+        print("[ERROR: failed to find package named '{0}']".format(pkg_name))
         sys.exit(-1)
 
     # Extract the first line of the full spec
@@ -471,7 +596,7 @@ class SpackEnv(UberEnv):
             spack_url = self.project_opts.get("spack_url", "https://github.com/spack/spack.git")
             spack_branch = self.project_opts.get("spack_branch", "develop")
 
-            clone_cmd =  "git {0} clone --single-branch --depth=1 -b {1} {2}".format(clone_opts, spack_branch, spack_url)
+            clone_cmd =  "git {0} clone --single-branch --depth=1 -b {1} {2} spack".format(clone_opts, spack_branch, spack_url)
             sexe(clone_cmd, echo=True)
 
         if "spack_commit" in self.project_opts:
@@ -480,7 +605,7 @@ class SpackEnv(UberEnv):
             sha1 = self.project_opts["spack_commit"]
             res, current_sha1 = sexe("git log -1 --pretty=%H", ret_output=True)
             if sha1 != current_sha1:
-                print("[info: using spack commit {}]".format(sha1))
+                print("[info: using spack commit {0}]".format(sha1))
                 sexe("git stash", echo=True)
                 sexe("git fetch --depth=1 origin {0}".format(sha1),echo=True)
                 res = sexe("git checkout {0}".format(sha1),echo=True)
@@ -489,11 +614,11 @@ class SpackEnv(UberEnv):
                     print("[ERROR: Git failed to checkout]")
                     sys.exit(-1)
 
-        if self.opts["spack_pull"]:
+        if self.opts["repo_pull"]:
             # do a pull to make sure we have the latest
             os.chdir(pjoin(self.dest_dir,"spack"))
             sexe("git stash", echo=True)
-            sexe("git pull", echo=True)
+            res = sexe("git pull", echo=True)
             if res != 0:
                 #Usually untracked files that would be overwritten
                 print("[ERROR: Git failed to pull]")
@@ -503,7 +628,7 @@ class SpackEnv(UberEnv):
         # disables all config scopes except "defaults", which we will
         # force our settings into
         spack_lib_config = pjoin(spack_dir,"lib","spack","spack","config.py")
-        print("[disabling config scope (except defaults) in: {}]".format(spack_lib_config))
+        print("[disabling config scope (except defaults) in: {0}]".format(spack_lib_config))
         cfg_script = open(spack_lib_config).read()
         for cfg_scope_stmt in ["('system', os.path.join(spack.paths.system_etc_path, 'spack')),",
                             "('site', os.path.join(spack.paths.etc_path, 'spack')),",
@@ -524,9 +649,9 @@ class SpackEnv(UberEnv):
 
         # copy in "defaults" config.yaml
         config_yaml = pabs(pjoin(cfg_dir,"..","config.yaml"))
-        sexe("cp {} {}/".format(config_yaml, spack_etc_defaults_dir), echo=True)
+        sexe("cp {0} {1}/".format(config_yaml, spack_etc_defaults_dir), echo=True)
         mirrors_yaml = pabs(pjoin(cfg_dir,"..","mirrors.yaml"))
-        sexe("cp {} {}/".format(mirrors_yaml, spack_etc_defaults_dir), echo=True)
+        sexe("cp {0} {1}/".format(mirrors_yaml, spack_etc_defaults_dir), echo=True)
 
         # copy in other settings per platform
         if not cfg_dir is None:
@@ -538,16 +663,16 @@ class SpackEnv(UberEnv):
             packages_yaml  = pjoin(cfg_dir,"packages.yaml")
 
             if os.path.isfile(config_yaml):
-                sexe("cp {} {}/".format(config_yaml , spack_etc_defaults_dir ), echo=True)
+                sexe("cp {0} {1}/".format(config_yaml , spack_etc_defaults_dir ), echo=True)
 
             if os.path.isfile(mirrors_yaml):
-                sexe("cp {} {}/".format(mirrors_yaml , spack_etc_defaults_dir ), echo=True)
+                sexe("cp {0} {1}/".format(mirrors_yaml , spack_etc_defaults_dir ), echo=True)
 
             if os.path.isfile(compilers_yaml):
-                sexe("cp {} {}/".format(compilers_yaml, spack_etc_defaults_dir ), echo=True)
+                sexe("cp {0} {1}/".format(compilers_yaml, spack_etc_defaults_dir ), echo=True)
 
             if os.path.isfile(packages_yaml):
-                sexe("cp {} {}/".format(packages_yaml, spack_etc_defaults_dir ), echo=True)
+                sexe("cp {0} {1}/".format(packages_yaml, spack_etc_defaults_dir ), echo=True)
         else:
             # let spack try to auto find compilers
             sexe("spack/bin/spack compiler find", echo=True)
@@ -600,7 +725,7 @@ class SpackEnv(UberEnv):
                     # testing that the path exists is mandatory until Spack team fixes
                     # https://github.com/spack/spack/issues/16329
                     if os.path.isdir(install_path):
-                        print("[Warning: {} {} has already been installed in {}]".format(self.pkg_name, self.opts["spec"],install_path))
+                        print("[Warning: {0} {1} has already been installed in {2}]".format(self.pkg_name, self.opts["spec"],install_path))
                         print("[Warning: Uberenv will proceed using this directory]".format(self.pkg_name))
                         self.use_install = True
 
@@ -615,9 +740,9 @@ class SpackEnv(UberEnv):
             if self.opts["ignore_ssl_errors"]:
                 install_cmd += "-k "
             if not self.opts["install"]:
-                install_cmd += "dev-build --quiet -d {} ".format(self.pkg_src_dir)
+                install_cmd += "dev-build --quiet -d {0} ".format(self.pkg_src_dir)
                 if self.pkg_final_phase:
-                    install_cmd += "-u {} ".format(self.pkg_final_phase)
+                    install_cmd += "-u {0} ".format(self.pkg_final_phase)
             else:
                 install_cmd += "install "
                 if self.opts["run_tests"]:
@@ -643,18 +768,22 @@ class SpackEnv(UberEnv):
                         break
                 if activate:
                     activate_cmd = "spack/bin/spack activate " + pkg_name
-                    sexe(activate_cmd, echo=True)
+                    res = sexe(activate_cmd, echo=True)
+                    if res != 0:
+                      return res
         # note: this assumes package extends python when +python
         # this may fail general cases
         if self.opts["install"] and "+python" in full_spec:
             activate_cmd = "spack/bin/spack activate /" + self.spec_hash
-            sexe(activate_cmd, echo=True)
+            res = sexe(activate_cmd, echo=True)
+            if res != 0:
+              return res
         # if user opt'd for an install, we want to symlink the final
         # install to an easy place:
         if self.opts["install"] or self.use_install:
             pkg_path = self.find_spack_pkg_path_from_hash(self.pkg_name, self.spec_hash)
             if self.pkg_name != pkg_path["name"]:
-                print("[ERROR: Could not find install of {}]".format(self.pkg_name))
+                print("[ERROR: Could not find install of {0}]".format(self.pkg_name))
                 return -1
             else:
                 # Symlink host-config file
@@ -665,24 +794,24 @@ class SpackEnv(UberEnv):
                     if os.path.islink(hc_fname):
                         os.unlink(hc_fname)
                     elif os.path.isfile(hc_fname):
-                        sexe("rm -f {}".format(hc_fname))
-                    print("[symlinking host config file to {}]".format(pjoin(self.dest_dir,hc_fname)))
+                        sexe("rm -f {0}".format(hc_fname))
+                    print("[symlinking host config file to {0}]".format(pjoin(self.dest_dir,hc_fname)))
                     os.symlink(hc_path,hc_fname)
 
                 # Symlink install directory
                 if self.opts["install"]:
-                    pkg_lnk_dir = "{}-install".format(self.pkg_name)
+                    pkg_lnk_dir = "{0}-install".format(self.pkg_name)
                     if os.path.islink(pkg_lnk_dir):
                         os.unlink(pkg_lnk_dir)
                     print("")
-                    print("[symlinking install to {}]".format(pjoin(self.dest_dir,pkg_lnk_dir)))
+                    print("[symlinking install to {0}]".format(pjoin(self.dest_dir,pkg_lnk_dir)))
                     os.symlink(pkg_path["path"],pabs(pkg_lnk_dir))
                     print("")
                     print("[install complete!]")
         # otherwise we are in the "only dependencies" case and the host-config
         # file has to be copied from the do-be-deleted spack-build dir.
         else:
-            pattern = "*{}.cmake".format(self.pkg_name)
+            pattern = "*{0}.cmake".format(self.pkg_name)
             build_dir = pjoin(self.pkg_src_dir,"spack-build")
             hc_glob = glob.glob(pjoin(build_dir,pattern))
             if len(hc_glob) > 0:
@@ -690,10 +819,10 @@ class SpackEnv(UberEnv):
                 hc_fname = os.path.split(hc_path)[1]
                 if os.path.islink(hc_fname):
                     os.unlink(hc_fname)
-                print("[copying host config file to {}]".format(pjoin(self.dest_dir,hc_fname)))
-                sexe("cp {} {}".format(hc_path,hc_fname))
-                print("[removing project build directory {}]".format(pjoin(build_dir)))
-                sexe("rm -rf {}".format(build_dir))
+                print("[copying host config file to {0}]".format(pjoin(self.dest_dir,hc_fname)))
+                sexe("cp {0} {1}".format(hc_path,hc_fname))
+                print("[removing project build directory {0}]".format(pjoin(build_dir)))
+                sexe("rm -rf {0}".format(build_dir))
 
     def get_mirror_path(self):
         mirror_path = self.opts["mirror"]
@@ -712,7 +841,7 @@ class SpackEnv(UberEnv):
         mirror_cmd = "spack/bin/spack "
         if self.opts["ignore_ssl_errors"]:
             mirror_cmd += "-k "
-        mirror_cmd += "mirror create -d {} --dependencies {}{}".format(mirror_path,
+        mirror_cmd += "mirror create -d {0} --dependencies {1}{2}".format(mirror_path,
                                                                     self.pkg_name,
                                                                     self.opts["spec"])
         return sexe(mirror_cmd, echo=True)
@@ -741,20 +870,20 @@ class SpackEnv(UberEnv):
 
         if existing_mirror_path and mirror_path != existing_mirror_path:
             # Existing mirror has different URL, error out
-            print("[removing existing spack mirror `{}` @ {}]".format(mirror_name,
-                                                                    existing_mirror_path))
+            print("[removing existing spack mirror `{0}` @ {1}]".format(mirror_name,
+                                                                        existing_mirror_path))
             #
             # Note: In this case, spack says it removes the mirror, but we still
             # get errors when we try to add a new one, sounds like a bug
             #
-            sexe("spack/bin/spack mirror remove --scope=defaults {} ".format(mirror_name),
+            sexe("spack/bin/spack mirror remove --scope=defaults {0} ".format(mirror_name),
                 echo=True)
             existing_mirror_path = None
         if not existing_mirror_path:
             # Add if not already there
-            sexe("spack/bin/spack mirror add --scope=defaults {} {}".format(
+            sexe("spack/bin/spack mirror add --scope=defaults {0} {1}".format(
                     mirror_name, mirror_path), echo=True)
-            print("[using mirror {}]".format(mirror_path))
+            print("[using mirror {0}]".format(mirror_path))
 
     def find_spack_upstream(self, upstream_name):
         """
@@ -795,8 +924,8 @@ class SpackEnv(UberEnv):
             sexe("rm spack/etc/spack/defaults/upstreams.yaml")
             with open('spack/etc/spack/defaults/upstreams.yaml','w+') as upstreams_cfg_file:
                 upstreams_cfg_file.write("upstreams:\n")
-                upstreams_cfg_file.write("  {}:\n".format(upstream_name))
-                upstreams_cfg_file.write("    install_tree: {}\n".format(upstream_path))
+                upstreams_cfg_file.write("  {0}:\n".format(upstream_name))
+                upstreams_cfg_file.write("    install_tree: {0}\n".format(upstream_path))
 
 
 def find_osx_sdks():
@@ -836,8 +965,8 @@ def setup_osx_sdk_env_vars():
 
     env["MACOSX_DEPLOYMENT_TARGET"] = dep_tgt
     env["SDKROOT"] = sdk_root
-    print("[setting MACOSX_DEPLOYMENT_TARGET to {}]".format(env["MACOSX_DEPLOYMENT_TARGET"]))
-    print("[setting SDKROOT to {}]".format(env[ "SDKROOT"]))
+    print("[setting MACOSX_DEPLOYMENT_TARGET to {0}]".format(env["MACOSX_DEPLOYMENT_TARGET"]))
+    print("[setting SDKROOT to {0}]".format(env[ "SDKROOT"]))
 
 
 
@@ -853,8 +982,8 @@ def main():
     # project options
     opts["project_json"] = find_project_config(opts)
 
-    # Initialize the environment
-    env = SpackEnv(opts, extra_opts)
+    # Initialize the environment -- use vcpkg on windows, spack otherwise
+    env = SpackEnv(opts, extra_opts) if not is_windows() else VcpkgEnv(opts, extra_opts)
 
     # Setup the necessary paths and directories
     env.setup_paths_and_dirs()
@@ -878,16 +1007,16 @@ def main():
     env.show_info()
 
 
-    ##########################################################
-    # we now have an instance of spack configured how we
-    # need it to build our tpls at this point there are two
-    # possible next steps:
+    ###########################################################
+    # we now have an instance of our package manager configured
+    # how we need it to build our tpls. At this point there are
+    # two possible next steps:
     #
     # *) create a mirror of the packages
     #   OR
     # *) build
     #
-    ##########################################################
+    ###########################################################
     if opts["create_mirror"]:
         return env.create_mirror()
     else:
@@ -903,3 +1032,4 @@ def main():
 
 if __name__ == "__main__":
     sys.exit(main())
+